pyproject.toml
[tool.poetry]
name = "fastapiwilly"
version = "0.1.0"
description = ""
authors = ["Leo Ramirez <leoramirez322@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.11"
uvicorn = "^0.22.0"
fastapi = "^0.95.1"
psycopg2-binary = "^2.9.6"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
alembic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
